backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
cyvcf2==0.31.4
pydantic==2.11.4
requests==2.32.3
google-genai>=1.0.0
pinecone>=3.0.0
python-dotenv>=1.0.0